Revert "jabagram/telegram: Fix issues with long timeouts between messages"
This issue seems to have been fixed in recent versions of aiohttp, with
a switch to using aiodns by default if available. In jabagram we already
have a dependency to aiodns because of slixmpp, so this should no longer
be an issue for us.

If you encounter the same issue again, please make sure you are using
the latest version of aiohttp.

This reverts commit f3ba7e54c663a8cf21f3d56729e6ceb3486ac8f4.