Introduce a system of actors to represent users from Telegram
Introduces a new entity called XmppActor - which actually encapsulates
connection to the XMPP server with another resource part and it is set
according to each user in Telegram. By default, at least one actor is
initialized to listen for events from XMPP (listener). The rest of the
XmppActor is initialized in a lazy way, and stored in a "pool" held by
the XmppActorFactory. The default size of the pool is 16 to minimize
overheads from multiple connections to server. However, this allows us
to greatly improve quality of communication across the bridge as it
eliminates the constant events of change of nicknames (which also
created garbage from events in clients like Gajim for example) and will
allow to implement in the future support for forwarding Telegram avatars
using XEP-0084.

Signed-off-by: Vasiliy Stelmachenok <ventureo@cachyos.org>